requirements.txt
streamlit
streamlit-lottie
langchain-openai
langchain-core
pydantic
python-dotenv
fpdf
pandas
requests